refactor: move contract creation struct to query api
